+++ packages/nextra/src/server/rehype-plugins/index.ts
@@ -6,7 +6,5 @@ export {
 export { rehypeBetterReactMathjax } from './rehype-better-react-mathjax.js'
 export { rehypeExtractTocContent } from './rehype-extract-toc-content.js'
 export { rehypeIcon } from './rehype-icon.js'
-export {
-  twoslashRenderer,
-  rehypeTwoslashPopup
-} from './twoslash/rehype-twoslash-popup.js'
+// We explicitly didn't put `rehype-twoslash-popup.ts` in index.js file to avoid
+// having warnings which comes from `@typescript/vfs` https://github.com/shuding/nextra/pull/3349
